Extracted book grid column related code snippets to BookGridColumn
