+++ src/main/java/com/karankumar/bookproject/ui/shelf/BookGrid.java
@@ -4,18 +4,13 @@ import com.karankumar.bookproject.backend.entity.Book;
 import com.karankumar.bookproject.backend.entity.PredefinedShelf;
 import com.karankumar.bookproject.backend.service.PredefinedShelfService;
 import com.karankumar.bookproject.ui.book.BookForm;
+import com.karankumar.bookproject.ui.shelf.component.BookGridColumn;
 import com.vaadin.flow.component.grid.Grid;
-import com.vaadin.flow.data.renderer.LocalDateRenderer;
 import lombok.extern.java.Log;
 
-import java.time.format.DateTimeFormatter;
-import java.time.format.FormatStyle;
-import java.util.Comparator;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
-
-import static com.karankumar.bookproject.ui.shelf.BooksInShelfView.*;
 
 
 @Log
@@ -28,16 +23,7 @@ public class BookGrid {
     }
 
     public void configure() {
-        resetGridColumns();
-
-        addTitleColumn();
-        addAuthorColumn();
-        bookGrid.addColumn(GENRE_KEY);
-        addDateStartedColumn();
-        addDateFinishedColumn();
-        bookGrid.addColumn(RATING_KEY);
-        bookGrid.addColumn(PAGES_KEY);
-        bookGrid.addColumn(PAGES_READ_KEY);
+        new BookGridColumn(bookGrid).addColumns();
     }
 
     void bind(BookForm bookForm) {
@@ -50,47 +36,6 @@ public class BookGrid {
             }
         });
     }
-
-    private void resetGridColumns() {
-        bookGrid.setColumns();
-    }
-
-    private void addDateFinishedColumn() {
-        bookGrid.addColumn(new LocalDateRenderer<>(Book::getDateFinishedReading, DateTimeFormatter.ofLocalizedDate(FormatStyle.MEDIUM)))
-                .setHeader("Date finished reading")
-                .setComparator(Comparator.comparing(Book::getDateStartedReading))
-                .setSortable(true)
-                .setKey(DATE_FINISHED_KEY);
-    }
-
-    private void addDateStartedColumn() {
-        bookGrid.addColumn(new LocalDateRenderer<>(Book::getDateStartedReading, DateTimeFormatter.ofLocalizedDate(FormatStyle.MEDIUM)))
-                .setHeader("Date started reading")
-                .setComparator(Comparator.comparing(Book::getDateStartedReading))
-                .setKey(DATE_STARTED_KEY);
-    }
-
-    private void addAuthorColumn() {
-        bookGrid.addColumn(AUTHOR_KEY)
-                .setComparator(Comparator.comparing(author -> author.getAuthor().toString()))
-                .setSortable(true);
-    }
-
-    private void addTitleColumn() {
-        bookGrid.addColumn(this::combineTitleAndSeries) // we want to display the series only if it is bigger than 0
-                .setHeader("Title")
-                .setKey(TITLE_KEY)
-                .setSortable(true);
-    }
-
-    private String combineTitleAndSeries(Book book) {
-        if (book.existsSeriesPosition()) {
-            return String.format("%s (#%d)", book.getTitle(), book.getSeriesPosition());
-        }
-
-        return book.getTitle();
-    }
-
 
     public void toggleColumnVisibility(String columnKey, boolean showColumn) {
         if (bookGrid.getColumnByKey(columnKey) == null) {
