+++ src/main/java/com/karankumar/bookproject/ui/shelf/BooksInShelfView.java
@@ -48,16 +48,6 @@ import static com.karankumar.bookproject.backend.entity.PredefinedShelf.ShelfNam
 @PageTitle("My Books | Book Project")
 @Log
 public class BooksInShelfView extends VerticalLayout {
-    // TODO: 3.08.2020 Encapsulate these variables. visibility should be package-access
-    public static final String TITLE_KEY = "title";
-    public static final String AUTHOR_KEY = "author";
-    public static final String GENRE_KEY = "genre";
-    public static final String PAGES_KEY = "numberOfPages";
-    public static final String PAGES_READ_KEY = "pagesRead";
-    public static final String RATING_KEY = "rating";
-    public static final String DATE_STARTED_KEY = "dateStartedReading";
-    public static final String DATE_FINISHED_KEY = "dateFinishedReading";
-
     public final BookGrid bookGrid;
     public final BookShelfComboBox whichShelf;
